Fix: Api 생성 및 service 계층 주입
Temp api 에서 실 api로 변경 및 service계층 주입

+++ src/main/java/com/example/demo/domain/gameRoom/entity/GameRoom.java
@@ -14,7 +14,7 @@ public class GameRoom {
     @Id
     @GeneratedValue
     @Column(name = "gameRoom_id")
-    private int id; //추후 long으로 변경 repository계층 구현후
+    private Long id; //추후 long으로 변경 repository계층 구현후
 
     //user 4명
